Improve type consistency

+++ Core/Hash.h
@@ -16,10 +16,11 @@ _force_inline bool isKeyEqual(const t_keyType &a, const t_keyType &b)
 	return a == b; 
 }
 
+
 template<>
 _force_inline bool isKeyEqual<const char*>(const char* const& a, const char* const& b)
 {
-	int i = 0;
+	i32 i = 0;
 	for (; a[i] != '\0'; i++) {
 		if (a[i] != b[i])
 			return false;
@@ -30,7 +31,7 @@ _force_inline bool isKeyEqual<const char*>(const char* const& a, const char* con
 
 
 template<typename t_keyType> 
-_force_inline size_t hash(const t_keyType &value)
+_force_inline u64 hash(const t_keyType &value)
 { 
 	return std::hash<t_keyType>{}(value); 
 }
@@ -45,7 +46,7 @@ _force_inline size_t hash::hash<const char *>(const char * const &value)
 
 	// Apply Fowler-Noll-Vo
 	age_assertFatal(value != nullptr, "Cannot hash a nullptr string");
-	size_t hash = k_seed;
+	u64 hash = k_seed;
 	for (u32 i = 0; value[i] != '\0'; i++) 
 		hash = (value[i] ^  hash) * k_prime;
